repurposed old tests for extra tests. Removed the excess.
